spring bean lifecycle

+++ src/main/java/com/logitech/domain/terminal/exception/TerminalNotFoundException.java
@@ -1,7 +1,7 @@
 package com.logitech.domain.terminal.exception;
 
 public class TerminalNotFoundException extends Exception{
-    public TerminalNotFoundException(int id) {
+    public TerminalNotFoundException(Long id) {
         super("Terminal with ID:: " + id + ", not found in database.");
     }
 }
